♻️ Refactor: Move queue header to different folder
