+++ client/src/features/queue/components/header/QueueHeader.tsx
@@ -1,6 +1,6 @@
-import styles from './Queue.module.scss';
+import styles from '../Queue.module.scss';
 import { RiCloseLargeFill } from 'react-icons/ri';
-import IconButton from '../../ui/IconButton/IconButton.tsx';
+import IconButton from '../../../../ui/IconButton/IconButton.tsx';
 
 interface QueueHeaderProps {
   onClick?: () => void;
